requirements.txt
pandas
requests
beautifulsoup4
openpyxl
streamlit
python-dateutil